[R-Debug, R-Console] Adapted to RJ-0.3.0

+++ de.walware.statet.r.core/srcDebug/de/walware/statet/r/internal/rdata/CombinedFactory.java
@@ -35,7 +35,7 @@ public class CombinedFactory extends RObjectFactoryImpl {
 	}
 	
 	public CombinedElement readObject(final ObjectInput in, final int flags, final CombinedElement parent, final RElementName name) throws IOException, ClassNotFoundException {
-		final int type = in.readInt();
+		final byte type = in.readByte();
 		switch (type) {
 		case -1:
 			return null;
