+++ de.walware.statet.r.core/srcDebug/de/walware/statet/r/internal/rdata/RVectorVar.java
@@ -95,7 +95,7 @@ public class RVectorVar<DataType extends RStore> extends CombinedElement
 	}
 	
 	
-	public int getRObjectType() {
+	public byte getRObjectType() {
 		return TYPE_VECTOR;
 	}
